refactor: move runtime helpers into runtime package
